set up App class for core logic and remove some circular dependencies
